feat: rework framestate logic and do not share tile queues
This attempts to clarify how OL frame state objects are handled: there is
one "root" frame state for a job, and all layers derive their frame state
from it. Each layer now also has a specific tile queue to avoid having
broken progress updates.
